Creates  src/test/verifyHeaders.
This commit creates the test file for the verify headers array.
There is a complie test, tests to verify the right errors are
thrown, tests to verify false is returned in the correct
secanrios, and a test to verify true is returned when the
proper information is given to the module.