Creates schema folder.
This commit creates the schema folder. This folder will be used
to keep information about different providers and the csv data.
Wether that be the headers on the csv data, notes about different
columns, and tags or ways to categorize things. This commit creates
5 different provier hashmaps which include the headers property
and each property on the headers is a column of the csv data for
the respective provider.