Updating src/lib/verifyHeaders and src/test/verifyHeaders.test.
This commit updates the verify headers module to be able to have
dependecy injection. This means the providers can be given as
a parameter instead of being hardcoded as the schema in the
project. This allows the module to be tested more efficently
and can be used with any supplied providers. The tests were
also upated to pass in object parameters. This commit updates
all the test so they pass properly.