Creates src/lib/toTransactions and src/test/toTransactions.test.
This commit creates the to transactions module. This module uses
dependency injection to allow the specification of the providers
schema. This module given a prodiver, json, and the providers
schema, will convert all the json into transaction objects and
return an array of the transaction objects. These can be used
to be inserted into the db or used to calculate totals. This
commit also tests the module to ensure the module complies,
throws the proper error, and works properly.