Creates src/lib/calculateTotals and src/test/calculateTotals.test.
This commit creates the calculate totals module which takes in
an array of transaction objects and returns the total amount
of debits categorized and the total amount of credits of the
transactions. This commit also creates the test file to make sure
the module complies and the totals are calculated properly.