Creates src/lib/bulkInsert and src/test/bulkInsert.test.
This commit creates the bulk insert module which can handles
inserting multiple records into a database. There is a mongo
function for this but the results of the inserts are not
given back along with not running pre-save events from the
model. This modules loops over the data provided and inserts
it into the database and returns all the objects that were
inserted. This commit also creates a complie test for the
module.