Creates src/app/finance-csv and src/test/finance-csv.test.
This commit creates the finance csv class. This is the main class
of the project. It allows dependencies to be injected into the
class. The methods are mostly all wrappers for the modules in
lib. This commit has a main mehtod which handles putting all the
lib modules together to process the raw csv. I left a todo to insert
the data into the spreadsheet. This commit also creates a complie
test but more testing needs to be done.