Creates src/lib/toSheetData.js.
This commit creates the to sheet data module. This module uses
the schema mapping of the totals to cells in order to create
an object that has the cell number pair as the key and the value
as the value to be inserted into that cell. This can be passsed
to the insert google sheet module to insert the data into the
google sheet.