Updating src/app/finance-csv - big update.
This commit updates a good chunck of the finance csv class. It
adds the toSheetData method and the insertGoogleSheet method
on the function. This way these can be invoked from an instance
of the a class. This commit also updates the main functionailty
to handling everything that is involved with a file. I put the
csv and sheet id in a playground file and it worked perfectly!